Don't set buildset files for non-live items ahead
I82848367bd6f191ec5ae5822a1f438070cde14e1 introduced an optimization to
set the buildset files for non-live items if they don't already have
files set.

Later Id533772f35ebbc76910398e0e0fa50a3abfceb52 changed the approach to
also build a layout for non-live items with config updates.

Unfortunately the initial code path for setting the buildset files for
non-live items ahead was still in place.

In combination with circular config updates, this can cause problems
when merges are not completed in the same order as requested, which can
definitely happen in a production system. So far this issue went
unnoticed in the tests as the order there seems to be deterministic.

This resolves the issue of multiple merge jobs for a bundle that
includes non-live changes arriving out of sequence by expecting each
merge job to only update its own queue item.

The reason Zuul was previously updating non-live queue items was due to
an optimization that tried to avoid spawning merge jobs for non-live
items, and instead applied a truncated files data structure to the items
ahead. This works because the files structure we get back from the
merger has the files for every change ahead, so we can just apply files
1..N to each change N in the queue.

This process was partially undone by a change which spawned merge
operations for non-live items if they have config updates.

What that change didn't take into account was that a bundle's items,
whether they are live or non-live, should always have the same file
contents.

This change resolves this by accepting that in the case of non-live
items, we will receive multiple merge job responses and it leaves it to
each item to only update its files. It appears that the only current use
of the BuildSet.files attribute is on build sets that where we are going
to generate a layout, which as described above and in the referenced
changes, should only be those buildsets for items that change the
configuration (and therefore will have their own merge job spawned).
This means that it should be safe to remove the code that set the files
attributes on the non-live items ahead.

Co-Authored-By: James E. Blair <jim@acmegating.com>
Change-Id: I38925e5fd0ed5ff45aab17d108740345716fd478