Document backing up and restoring project keys
Administrators may not realize until it's too late that it would
have been better for them to save copies of project keys in case of
a disaster impacting the keystore in ZooKeeper. Zuul provides the
tools necessary to facilitate this, so document them prominently in
the places newcomers are likely to look when first installing or
learning to operate the software.

Change-Id: Iaf8bc9cbcdb885cd6b9648ec4080c144d1fb4ccf