Update patch set 1
Patch Set 1:

The commit message has great background, but could use more elaboration on exactly what this change is doing and why that's okay.  Here's what I wrote to try to explain it:

This resolves the issue of multiple merge jobs for a bundle that
includes non-live changes arriving out of sequence by expecting
each merge job to only update its own queue item.

The reason Zuul was previously updating non-live queue items was
due to an optimization that tried to avoid spawning merge jobs for
non-live items, and instead applied a truncated files data structure
to the items ahead.  This works because the files structure we get
back from the merger has the files for every change ahead, so we can
just apply files 1..N to each change N in the queue.

This process was partially undone by a change which spawned merge
operations for non-live items if they have config updates.

What that change didn't take into account was that a bundle's items,
whether they are live or non-live, should always have the same file
contents.

This change resolves this by accepting that in the case of non-live
items, we will receive multiple merge job responses and it leaves it
to each item to only update its files.  It appears that the only
current use of the BuildSet.files attribute is on build sets that
where we are going to generate a layout, which as described above and
in the referenced changes, should only be those buildsets for items
that change the configuration (and therefore will have their own merge
job spawned).  This means that it should be safe to remove the code
that set the files attributes on the non-live items ahead.

Patch-set: 1
CC: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Attention: {"person_ident":"Gerrit User 27582 \u003c27582@4a232e18-c5a9-48ee-94c0-e04e7cca6543\u003e","operation":"ADD","reason":"Someone else replied on the change"}