Use os.open with setpag
When we open the ioctl file to run the openafs setpag syscall,
we previously used the high-level open method, which apparently
issues an unwanted TCGETS ioctl which crashes the program with
a kernel error under certain versions of python+openafs+linux
(3.10.6, 1.8.9, 5.15.0, respectively).

Switch to a low-level open to avoid this call.

Change-Id: I5e08a6020cf6cd4ad2a0084effb697aa39dae9c6