Add manager/reporter support for config warnings
We recently added a severity field to configuration errors (but all
errors are currently at the "error" severity).  To prepare for
"warning" severity, update the pipeline managers and reporters to
expect both warnings and errors.

Errors will still trigger buildset failures, but warnings will not.
Both will be reported as comments.

Change-Id: Ia24e91f5ddff7d9869e9e83886f996e4f425e110