Attach loading_errors to the ParseContext
We pass around a single LoadingErrors object in quite a bit of the
config parser, and upcoming changes will pass it around even more.
Most of those instances also have a ParseContext as well, and the
two objects have the same lifecycle and scope, so let's attach the
LoadingErrors to the ParseContext and simplify things.

Change-Id: I57d0edc3c69eace9ba12c2975278a1b9312f9d67