job: Allow dict in file matchers
This change allows you to specify a dict element in the job file
matcher; for example

  job:
    name: Parent job
    files:
     - foo
     - { name : 'bar' }
     - { name : ['foo', 'goo'] }

would all be valid for files/irrelevant-files.

It's a peculiarity of YAML that you can insert the keys of one map
into another map [1], but you can't extend a list.  Having the element
in dict format allows you to do some grouping and merging across jobs
in YAML, e.g.

  job:
    files:
      - &A_FILES { name : ['foo', 'bar' ] }
      - &B_FILES { name : ['moo', 'goo' ] }

  job:
    files:
      - blah
      - << : *A_FILES

would be valid to include that group of files into your job matcher.

We've had some discussion on an "always-files" mechanism to do
something like this with job inheritance in
Iec596b9f967d5a6080874e9e438db3ce8bd76e5c.  This strikes me as being a
different way to achieve something similar outside having to have
strict job inheritance.  I think both could be useful in different
circumstances.

YAML's behaviour here is admittedly quirky, but if nothing else there
are examples of it.  This was inspired by a similar construct used in
opendev to use common node definitions [2].

[1] https://yaml.org/type/merge.html
[2] https://opendev.org/opendev/system-config/src/commit/6a09102d82d595739e605cd3581a04ae130daba0/zuul.d/system-config-run.yaml#L98

Change-Id: Id4a0e9028305624a89c6bdb1c6960de9de5f9601