[WIP] Add dict to file matchers to allow for merging
This change allows you to specify a dict to the job file matcher,
which allows you to use YAML merging to inherit common files you would
like to match.  i.e.

  job:
    name: Parent job
    files:
     - &BASE_FILES { files: ['moo', 'goo'] }

  job:
    name: Another job
    files:
      - << : *BASE_FILES
      - foo
      - bar

It's a peculiarity of YAML that you can insert the keys of one map
into another map, but you can't extend a list.  Above, "another job"
would be transformed by YAML parsing into essentially

  job:
    name: Another job
    files:
      - files:
          - moo
          - goo
      - foo
      - bar

This doesn't require much updating to the parsing routine; it just
needs to check if the file matcher it is parsing is a dict, and if so
then walk into it's "files" key.

We've had some discussion on an "always-files" mechanism to do
something like this with job inheritance in
Iec596b9f967d5a6080874e9e438db3ce8bd76e5c.  This strikes me as being a
fairly good way to achieve the same sort of thing (define common file
matchers once) without too much change to job semantics.

This was inspired by a similar construct used in opendev to use common
node definitions [2].

[1] https://yaml.org/type/merge.html
[2] https://opendev.org/opendev/system-config/src/commit/6a09102d82d595739e605cd3581a04ae130daba0/zuul.d/system-config-run.yaml#L98

Change-Id: Id4a0e9028305624a89c6bdb1c6960de9de5f9601