Use built-in URL data type instead of custom parse
The getHomepageUrl function searches for /# when trying to find hash
anchors in the URL. This misses URLs that don't end with a trailing slash
(i.e http://localhost:3000/auth_callback#state=231231231)

To fix this, this commit switches to using the built-in URL data type
which handles this case and all the other cases. It parses an URL and
returns its origin by accessing the `origin` attribute.

A trailing slash is added to keep backwards compatibility as it is
expected everywhere in the codebase.

Change-Id: I15827f4cb4c28f3163a4b4e84b872331b972b156
Signed-off-by: Flavio Percoco <flavio@pacerevenue.com>