Add explicit foreign key indexes
The artifact, provides, and build_event tables all have foreign key
references to the build table.  In MySQL this causes an automatic
index to be created, but that does not appear to be the case in
Postgres.  Without these indexes, many of the queries we do can be
quite costly.

To address this, explicitly create these indexes.  Under MySQL, this
will simply rename the existing indexes, so this should be safe
and effective for both systems.

Change-Id: I16223fba75c1295480431d03ac59f72a5281b498