Update patch set 8
Patch Set 8:

I'm trying to say that when I consulted the whitelabel site preview, it did not function as expected; it changed the url in the browser in a way that the current whitelabel build does not.  That suggests an error in the change.

You can see the same thing if you look at the build right now.

But if you want to compare it to the current behavior, there are no other whitelabel builds available because their logs have all expired.  So you'll need to propose a no-op change that triggers a whitelabel build in order to see the current behavior.  Prefix the subject with "DNM" so people know the change is not meant for merging.

Patch-set: 8
Attention: {"person_ident":"Gerrit User 6159 \u003c6159@4a232e18-c5a9-48ee-94c0-e04e7cca6543\u003e","operation":"ADD","reason":"Someone else replied on the change"}