Refactor Gerrit driver event sources
Gerrit supports a number of pub-sub plugins which can act as
alternatives to stream-events.  These can often be easier for
users to configure than ssh access and have the advantage of
providing queueing and delivery guarantees for messages.

Subsequent changes will add support for multiple pub-sub event
sources, so to make driver maintenance easier, this change
refactors the gerrit driver into its two current event sources:
SSH stream-events and the checks plugin.

The checks plugin is a bit of a special case in that we always
start the polling method for it, so event source selection isn't
quite as clean as for the other sources.  But it's still useful
to compartmentalize it as much as possible, so it is moved to
its own file and treated as similarly as possible.

The stream events listener behaves much more like the pub-sub
listeners will.

Change-Id: I28d0f1e37b87927c5f2dd5e9bdc162391ad66d07