WIP: Add Kafka support to Gerrit
Gerrit supports a number of pub-sub plugins which can act as
alternatives to stream-events.  These can often be easier for
users to configure than ssh access and have the advantage of
providing queueing and delivery guarantees for messages.

This change not only adds support for Kafka, but is meant as
a template for adding support for other Gerrit pub-sub plugins
as well.

Change-Id: Ib03d8cb9ef709b625d2717a09125930548c86a22