Add commit_id to zuul vars
We added the commit_id to the MQTT reporter a while ago so that
external systems which identify changes by commit id (the Gerrit
current patchset hexsha, or the Github PR HEAD, etc) can track
Zuul work.  To enable the same sort of interfacing with external
systems in Zuul jobs, the variable is added to the set of zuul
variables.

To make it easier for use in multiple pipelines, the variable is
set to oldrev or newrev in the non-change case.

To make it easier to confirm driver parity, a new Gerrit driver
test class is added with the intention that it should parallel
the other driver test classes over time.

Change-Id: I2826349ee9ce696c8f9de8c23276094123292196