Change start_time and end_time field type to integer; add new fields
The start_time and end_time fields are a unix timestamp value, which is
always integer. By changing that field type to integer, the Grafana
OpenSearch plugin can properly parse the datetime.
Also added in that change new fields to the document: start_datetime and
end_datetime which represents conversion from timestam to datetime in
ISO format. This might be a better solution for those, who are using
the Grafana Opensearch plugin to create dashboard with the build
statistics.

Change-Id: I95b043699ed390dffdf50a9463f1bc2e81017f04