Fix null dereference in gitlab
The pipeline manager may call updateChange to resolve changes
that are not in the change cache.  In these cases, there is no
associated event object available.  However, the gitlab driver
incorrectly assumed that it would always be available and attempted
to populate change fields from the event.  This could cause
pipeline processing to become permanently stuck.

The intent is that updateChange should simply update information
about the change from the source.  It shouldn't really need to use
information from the event at all.  But because gitlab made some
data population asynchronous, the merge request info may not have
the head sha at the moment that the merge request was created, but
the event does.

In short, we should be able to get the information we need by:

1) looking for it in the MR in case the async update is done
2) getting it from the event if the event is present
3) Set it to None in all other cases

This way we try our best to populate information about the change,
and are likely to succeed most of the time.  If we do fail, we
won't be trying to dereference an event object that is None.

Change-Id: I1e6c94adaab989c8fc2c88912dc4000bd2a2eea1