Refactor will_retry
This moves some variables around to try to make the code have more
of a direct correspondence with the zuul_will_retry Ansible variable.

The logic is just a little too complex to have a 1:1 match between
a python variable and Ansible.  But this refactor adds a "should_retry"
which is pretty close to "zuul_will_retry" with the caveat that
will_retry is should_retry iff we have not hit the retry limit.

Change-Id: Ic869e3befc5763ffcd4a97c406224f7bb46acf13