Allow executor shutdown when disconnected from ZK
If we lose our connection to ZK and then receive a request to
gracefully stop, the stop will fail (even with no jobs running)
because we are unable to update the component state in ZK.

Make these updates a "best effort" and continue with the stop
procedure regardless.  This could aid operators who are trying
to stop a malfunctioning system.

Change-Id: I0d14ae9f8200cac12a02fa1ab5716c1f22863406