Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Weak -1: Can we update a test to check for this?  There's probably a test that exercises this already that we can just add an assertion for the number of merge jobs to.  I ask because this seems like something that we could easily regress and not notice.

If it's super hard to add a test, I can be talked around to a +2 as it otherwise lgtm.

Patch-set: 2
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1, 41ab6b34e99b1c61599358befd9452469e87c9f4