Update patch set 4
Patch Set 4:

PS4 was an experiment with reducing the scope of the unparsed abide lock, but it has problems that I don't think are worth solving.  I've restored PS3 which has a coarse unparsed abide lock during startup and reconfiguration, but still doesn't require it during tenant layout update, which is the main benefit.

Patch-set: 4