Sort job's `affected_projects` data attribute
This change addesses an issue with job deduplication where two jobs did
not correctly compared equal due on two schedulers due to a different
order of elements in the `affected_projects` job attribute.

The list of `affected_projects` of a job is constructed using a set to
eliminate duplicates and then converted to a list. Since sets are
unordered we need to return a sorted list instead of simply converting
the set to a list.

Change-Id: I89693c7badae7993db621ba8ae2fe4d2162829e4