Use sorted list for index of deduplicated secrets
When freezing a job we iterate over a set of (data) attributes which
also includes the playbooks. In this loop we freeze the playbooks
and so far also deduplicated the secrets.

Since the order in which we process those attributes is
non-deterministic we could also end up with a list of secrets that was
ordered differently for the same job on different schedulers. As the
index is used as a reference in the frozen playboooks this could break
job deduplication as the jobs no longer compared equal.

To fix this we will deduplicate the secrets after we've processed all
playbooks and use and store a list of secrets sorted by the secret name.

Change-Id: Ib84bfedb6740df73530aed7d82b8234a07757c59