Document max_threads_per_installation
This setting is proving to be useful for GitHub Enterprise users.

Because of the abuse detection described at [1], I am not comfortable
setting this to >1 for all GitHub users (especially github.com users),
and it may even still be reasonable to keep it at 1 for small GHE
installations.  In other words, while I would like to avoid adding
tuning parameters like this, I think this one is too dependent on
local external conditions to develop a reasonable heuristic, so we
should make it a first-class option.

[1] https://docs.github.com/en/rest/guides/best-practices-for-integrators?apiVersion=2022-11-28#dealing-with-rate-limits

Change-Id: I093c9958ed287f0c3ae0aff50c73299ecafa4627