Ensure that bundle is set correctly on items outside of change queue
This should fix a bug where a pipeline got stuck because of bundle that
wasn't deserialized properly:

    2023-08-18 06:32:05,194 ERROR zuul.Scheduler: Exception in pipeline processing:
    Traceback (most recent call last):
      File "/opt/zuul/lib/python3.11/site-packages/zuul/scheduler.py", line 2357, in _process_pipeline
        while not self._stopped and pipeline.manager.processQueue():
                                    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
      File "/opt/zuul/lib/python3.11/site-packages/zuul/manager/__init__.py", line 1797, in processQueue
        item_changed, nnfi = self._processOneItem(
                             ^^^^^^^^^^^^^^^^^^^^^
      File "/opt/zuul/lib/python3.11/site-packages/zuul/manager/__init__.py", line 1655, in _processOneItem
        ready = self.prepareItem(item)
                ^^^^^^^^^^^^^^^^^^^^^^
      File "/opt/zuul/lib/python3.11/site-packages/zuul/manager/__init__.py", line 1448, in prepareItem
        build_set.setConfiguration(self.current_context)
      File "/opt/zuul/lib/python3.11/site-packages/zuul/model.py", line 4615, in setConfiguration
        self.dependent_changes = [self._toChangeDict(i) for i in items]
                                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
      File "/opt/zuul/lib/python3.11/site-packages/zuul/model.py", line 4615, in <listcomp>
        self.dependent_changes = [self._toChangeDict(i) for i in items]
                                  ^^^^^^^^^^^^^^^^^^^^^
      File "/opt/zuul/lib/python3.11/site-packages/zuul/model.py", line 4625, in _toChangeDict
        change_dict['bundle_id'] = item.bundle.uuid
                                   ^^^^^^^^^^^^^^^^
    AttributeError: 'dict' object has no attribute 'uuid'

Usually a bundle should never be a dict outside of ZooKeeper. The only
reason a bundle might not be deserialized properly is if the item that
references the bundle in the code above is not part of the queue
anymore. In this case the item won't be deserialized within the
deserialization() method of the ChangeQueue.

The only place left where a QueueItem is deserialized is within the
deserialization() method of the Bundle itself. In this case we only
deserialize the QueueItem but don't add any bundle information to it.

In order to fix this, we can directly provide the bundle to the
QueueItem when deserializing it in the Bundle class. Additionally, we
add a check to ensure that the item is still in ZooKeeper. Otherwise the
queue processor might fail later on when it finds out that the item is
not part of ZooKeeper anymore.

Change-Id: Ie0c97cd5294d30523273a55ec5438f01cd37988e