Switch from fb-re2 to google-re2
At the time we began to adopt the re2 library, fb-re2 was the only
Python wrapper.  Google-re2 is available now, is developed and
released by the maintainers of re2 as part of the re2 project,
and distributes manylinux wheels which simplifies the Zuul build
process.  We have already adopted google-re2 for our Ansible
environments, so this simplifies packaging.

Change-Id: Id1b47d3b980faaf48a3d2b7959925d8c73bd3db1