Make Ansible variable freezing more efficient
We currently iterate over every job/host/etc variable in the freeze
playbook.  The reason is because if any value in any variable is
Undefined according to jinja, the Ansible combine filter throws
an error.  What we want to do in Zuul is merge any variable we can,
but if any is undefined, we skip it.  Thus, the process of combining
the variables one at a time in a task and ignoring errors.

This process can be slow, especially if we have start with a large
amount of data in one of the early variables.  The combine filter
needs to reprocess the large data repeatedly for each additional
variable.

To improve the process, we create a new filter, "zuul_combine" which
is a simplified version of the Ansible combine filter.  It only
operates on two dictionaries, doesn't offer any options for controlling
merging, and ignores errors when setting individual variables (while
still checking for Undefined values and skipping them).

Change-Id: Ie41f404762daa1b1a5ae47f6ec1aa1954ad36a39