Update patch set 1
Patch Set 1:

I tried a few different ways of representing this, and this seemed the most compact and intuitive (despite it potentially looking like the window is actually a max -- it isn't really, but it sort of is -- at least, it's the max number of changes to run jobs for at a time -- and that seemed semantically close enough that it might be okay to represent this way).  I didn't want to spell it out on the same level as the queue name because we already have a lot of words on the page and that seemed to take up too much real estate.  But I'm open to that or other ideas if folks would prefer something else.

Patch-set: 1