Update patch set 3
Patch Set 3:

I'm not sure the projects page is the best fit for this.  Usually in the web ui we show information that is actionable for users.  But if there is some problem with a connection, that's probably something that an administrator needs to fix.  If there's an error with a connection, it should be logged at error level, and hopefully administrators know to look for those errors and act on them.

However, it might be a good idea to exponse information like this as a monitoring metric.  As a statsd and/or prometheus value, it can be tracked over time and used for alerts.

Patch-set: 3
CC: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>