Add more safety check tests
These tests would fail without the safety check introduced in the
parent change.

They were originally written under the assumption that we would
have the optimal behavior for the pipeline processor, where some
builds may be canceled but others may proceed after certain updates
to circular dependency bundles.  However, fixing that is proving
impractical without a refactor, so these tests are added in updated
form which mostly asserts that a lot of jobs are aborted.

The original form of the tests are also added in this change, but
with skip decorators attached.  It is hoped that after some
refactoring of circular dependency handling, we can use the original
form of these tests to validate the desired behavior.

Change-Id: Ie45da9fc1848a717bf5308e595edd27e598d6882